More control on showcase widget
With these changes, the Showcase widget has more control over ShowCaseWidget, and has some new ease of use changes.

>The key parameter is now nullable, when the key parameter is null the child is returned instead

This is useful when using the showcase widget inside ListView.builder and other generated widgets. Of coursei performance should be tested.

>New Showcase parameter - disableBarrierInteraction

This enables the Showcase widget to disable barrier clicks regardless of ShowCaseWidget settings

>New Showcase parameter - disableShowCaseWidgetOverlayChildren

This parameter allows the Showcase widget to disable the overlayChildren of the ShowCaseWidget for this Showcase. The usecase on mind with this change is to allow the developer to disable skip/next buttons and allow the user to click on the widget to proceed.